Apresentação dos resultados.

+++ calclactase/src/business/CalculatorBusiness.js
@@ -1,9 +1,113 @@
 import ProductBusiness from "./ProductBusiness";
+import SettingsBusiness from "./SettingsBusiness";
 
 let items = [];
 
+const FccAmountToHydrolyzeAGramOfLactose = 300;
+
 export default class CalculatorBusiness {
-  static getAll() {
+  static getResult = (medicineTypeId, medicineFcc) => {
+    console.log(">>> BEGIN getResult");
+    const result = {
+      resultItems: [],
+      lactaseSum: 0,
+      necessaryFcc: 0,
+      necessaryMedicineAmount: 0,
+      medicineDisplayName: "",
+    };
+
+    if (items.length === 0 || !medicineTypeId || !medicineFcc) return result;
+
+    result.resultItems = this.getResultItems();
+    result.lactaseSum = this.getLactoseSum(result.resultItems);
+    result.necessaryFcc = this.getNecessaryFcc(result.lactaseSum);
+    result.necessaryMedicineAmount = this.calcNecessaryMedicineAmount(
+      result.necessaryFcc,
+      medicineFcc
+    );
+    result.medicineDisplayName = this.getMedicineDisplayName(medicineTypeId, result.necessaryMedicineAmount);
+    console.log(">>> END getResult");
+    return result;
+  };
+
+  static getMedicineDisplayName = (medicineTypeId, necessaryMedicineAmount) => {
+    const medicineType = SettingsBusiness.getMedicineTypeById(medicineTypeId);
+
+    return necessaryMedicineAmount > 1
+      ? medicineType.pluralName
+      : medicineType.name;
+  };
+
+  //#region Lactose Calcs
+  static calcNecessaryMedicineAmount = (necessaryFcc, medicineFcc) => {
+    let necessaryMedicineAmount = necessaryFcc / medicineFcc;
+
+    let integerPartOfNecessaryMedicineAmount = Math.trunc(
+      necessaryMedicineAmount
+    );
+
+    let decimalPartOfNecessaryMedicineAmount = Number(
+      (necessaryMedicineAmount - integerPartOfNecessaryMedicineAmount).toFixed(
+        4
+      )
+    );
+
+    let finalNecessaryMedicineAmount =
+      decimalPartOfNecessaryMedicineAmount < 0.5
+        ? integerPartOfNecessaryMedicineAmount
+        : integerPartOfNecessaryMedicineAmount + 0.5;
+
+    if (finalNecessaryMedicineAmount < 0.5) finalNecessaryMedicineAmount = 0.5;
+
+    console.log("necessaryMedicineAmount: ", necessaryMedicineAmount);
+    console.log(
+      "decimalPartOfNecessaryMedicineAmount: ",
+      decimalPartOfNecessaryMedicineAmount
+    );
+
+    return finalNecessaryMedicineAmount;
+  };
+
+  static getLactoseSum = (resultItems) =>
+    resultItems.reduce((sum, item) => sum + item.lactoseValue, 0);
+
+  static getNecessaryFcc = (gramsOfLactose) =>
+    gramsOfLactose * FccAmountToHydrolyzeAGramOfLactose;
+
+  static getResultItems = () => {
+    console.log(">>> getResultItems");
+
+    const resultItems = [];
+
+    items.map((item) => {
+      const product = ProductBusiness.getById(item.productId);
+
+      const resultItem = {
+        itemId: item.id,
+        productId: product.id,
+        productName: product.name,
+        finalUnitValue: product.displayUnitValue * item.quantity,
+        displayUnit: product.displayUnit,
+        imageUrl: product.imageUrl,
+        quantity: item.quantity,
+        lactoseValue: this.calcLactoseValue(
+          item.quantity,
+          product.grams,
+          product.percentageOfLactose
+        ),
+      };
+
+      resultItems.push(resultItem);
+    });
+
+    return resultItems;
+  };
+
+  static calcLactoseValue = (quantity, weight, lactosePercentage) =>
+    quantity * weight * lactosePercentage;
+  //#endregion
+
+  static getAllItems() {
     return items;
   }
 
@@ -12,6 +116,7 @@ export default class CalculatorBusiness {
   };
 
   static Add = (productId) => {
+    console.log(">>> BEGIN CalculatorBusiness.Add");
     const itemIndex = this.findIndexByProductId(productId);
     console.log("Quantidade de itens: ", items.length);
 
@@ -23,11 +128,12 @@ export default class CalculatorBusiness {
 
     let item = {
       id: Math.random().toString(),
-      product: ProductBusiness.getById(productId),
+      productId: productId,
       quantity: 1,
     };
 
     items.push(item);
+    console.log(">>> END CalculatorBusiness.Add");
   };
 
   static addQuantity = (id) => {
@@ -38,11 +144,8 @@ export default class CalculatorBusiness {
   static removeQuantity = (id) => {
     const item = this.findById(id);
     item.quantity -= 1;
-    
-    if (item.quantity === 0) {
-      this.deleteById(id);
-      return;
-    }
+
+    if (item.quantity === 0) this.deleteById(id);
   };
 
   static findById = (id) => items.find((i) => i.id === id);
@@ -50,9 +153,9 @@ export default class CalculatorBusiness {
   static findIndexById = (id) => items.findIndex((i) => i.id === id);
 
   static findIndexByProductId = (productId) =>
-    items.findIndex((i) => i.product.id === productId);
+    items.findIndex((i) => i.productId === productId);
 
   static deleteById = (id) => {
-    items = items.filter(i => i.id !== id);
-  }
+    items = items.filter((i) => i.id !== id);
+  };
 }
